first attempt at rfsoc implementation methods in tcpborphserver
Adds the first pass implementation of supporting rfsoc in tcpborphserver. Right
now the rfsoc support is added within the default `raw` mode of tcpborphserver.
Going through the code I understood there to be different `modes` or
`personalities` that tcpborphserver coudl enable and in the future this seems
like something we probably should do but since it doesn't look like that is used
very often I didn't want this to be the first thing to test this capability at
the time.

The supported implementation now is to insert the initialization of the rfdc
driver which uses the linux device tree to initialize the driver. With the
driver initialized this also simplifies the extension of using the driver for
other method such as getting rfdc status and accessing more advanced features
like multi-tile synch and nco updates. The `rfdc_status_cmd` gives a little bit
of an idea how this would be done to other methods.

In addition this supports programming onboard rf plls (typically onboard LMK/LMX
PLLs to drive required RFDC clocks) for the ZCU216 and ZRF16 boards (not yet
tested on ZRF). The structure for supporting other boards is shown how it could
be extended but this does pose a problem of becoming large with supporting
multiple boards.

The programming of the rfplls is done through linux i2c driver and additionally
has a method for uploading the TICs register format for LMK/LMXs. Right now, a
custom `rfdc_upload_rfclk_cmd` is added patterend off much of what other methods
(like saveremote) in `upload.c` does. It was chosen just for the casper demo to
just use this method to have something ready but am needing to circle back to
decide on keeping this method or using what `upload.c` already provides.