modify dto application logic, rfsoc i2c clk platforms
The dto was applied in the rfsoc init command. This had a few
flaws in the approach. The dto is now instead managed as a seperate command.
Additionally, the logic now assumes an already uploaded dto at the bof base
directory by the name `tcpborphserver.dtbo`.

This implementation now provides an option to `apply` or `remove` the dto. When
apply is given with an active `dto` it will first be removed before applying the
new one. This differs from the previous implementation that would exit if one
was already applied and not previously removed.

It is important to note that the ideal integration of the device tree overlay
into tbs would be that the application of the overlay would also program the PL
using a fragment in the dto that also targets the fpga manager. Instead, as tbs
already uses the fpga manager to program the PL with the `.fpg` framework using
the dto approach would require a little bit of a rework of that logic since
using the dto would be enough to kick of pl programming.