update rfdc driver version
Move from rfdc driver v8.0 to v8.1. This requires versions of required libraries
(e.g., libmetal) attached with the Vitis 2020.2 sysroot. Or, one could go and
get the libraries from source and build to the correct versions. In this case,
just checked out the 2020.2 version of the rfdc driver from embedded software
and compiled against the 2020.2 sysroot and copied librfdc.so.1.1 to the rfsoc.

The previous driver version was limited to only supporting capabilities in the
rfdc up through Gen 3 ES silicon devices. Targeting production gen 3
devices with the previous version was possible but it would always limit
capabilities to that of ES1. For example, DSA would be capped at 11 dB. There
would also be some non-standard initialization of the DSA to -2 dB. This was
caused by the driver.

Rework the get/set dsa command; Add the VOP and get output current commands;
Modified to a more uniform response output for clients to reliably parse return
info; fixed a few bugs and edge cases in checking against number of
tiles/blocks.

Additionally added the get block status for some other testing that needed to be
done.