modify/remove mts command metods
Had kept getting the dac decimation factor even reporting information for dac
mts. Correctly get either the decimation or interpolation factor depending on
the converter type.

The detailed marker report for mts is no longer given as output when running
mts. The `?rfdc-mts-report` command is now `?rfdc-mts-debug-info` and must be
called instead to get that detailed marker report.

`?rfdc-report-mts-latency` has been removed. It was never formatted correctly to
parse responses in a clean way and is why `?rfdc-get-mts-tile-latency` was added.
The same information can still be retrieved in a more actionable way now with
`?rfdc-get-mts-tile-latency` by iterating over the enabled tiles in the mts mask.